Add null checks to autostart, finish function docs

+++ swm/src/autostart.c
@@ -14,6 +14,10 @@ Autostart Autostart__new(void)
 
 void Autostart__add(Autostart *as, const char *cmd)
 {
+	if (!as || !cmd) {
+		return;
+	}
+
 	if (as->size < MAX_AUTOSTART_COMMANDS + 1) {
 		as->cmds[as->size++] = (char *)cmd;
 		as->size += 1;
@@ -24,6 +28,10 @@ void Autostart__add(Autostart *as, const char *cmd)
 /// Function to execute shell commands and store their PIDS
 void Autostart__exec(Autostart *as)
 {
+	if (!as) {
+		return;
+	}
+
 	char buffer[MAX_AUTOSTART_COMMANDS] = { 0 };
 
 	for (uint32_t i = 0; i < as->size; ++i) {
@@ -46,6 +54,10 @@ void Autostart__exec(Autostart *as)
 /// them.
 void Autostart__kill(Autostart *as)
 {
+	if (!as) {
+		return;
+	}
+
 	for (uint32_t i = 0; i < as->size; ++i) {
 		if (as->pids[i] > 0) {
 			kill(as->pids[i], SIGTERM);
